chess_backend_bishal/requirements.txt
drf-yasg
pyotp
django-cors-headers
